requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
pytz>=2023.3
aiosqlite>=0.19.0
MetaTrader5>=5.0.45
openai>=1.0.0